move filter factory to package
